+++ src/lib/pages/proposal-details/components/ResultExplanation.tsx
@@ -11,16 +11,13 @@ import type {
   ProposalData,
   ProposalParams,
   ProposalVotesInfo,
+  Ratio,
   Token,
   TokenWithValue,
   U,
 } from "lib/types";
 import { ProposalStatus } from "lib/types";
-import {
-  divWithDefault,
-  formatPrettyPercent,
-  formatTokenWithValueList,
-} from "lib/utils";
+import { formatPrettyPercent, formatTokenWithValueList } from "lib/utils";
 
 import { ErrorFetchingProposalInfos } from "./ErrorFetchingProposalInfos";
 
@@ -83,10 +80,8 @@ export const ResultExplanation = ({
     params,
     proposalData.isExpedited
   );
-  const { yes, noWithVeto, nonAbstainVotes, totalVotes } =
+  const { totalVotes, yesNonRatio, noWithVetoRatio } =
     normalizeVotesInfo(votesInfo);
-  const yesRatio = divWithDefault(yes, nonAbstainVotes, 0);
-  const noWithVetoRatio = divWithDefault(noWithVeto, totalVotes, 0);
 
   if (proposalData.status === ProposalStatus.DEPOSIT_PERIOD) {
     const required = mapDeposit(proposalData.totalDeposit, minDeposit).reduce<
@@ -111,7 +106,7 @@ export const ResultExplanation = ({
   }
 
   if (proposalData.status === ProposalStatus.VOTING_PERIOD) {
-    if (totalVotes.lt(quorum))
+    if (totalVotes < quorum)
       return (
         <Text variant="body2">
           As of now, the proposal has not yet reached the required quorum. If
@@ -120,7 +115,7 @@ export const ResultExplanation = ({
         </Text>
       );
 
-    if (noWithVetoRatio.gte(vetoThreshold))
+    if (noWithVetoRatio >= vetoThreshold)
       return (
         <Text variant="body2">
           The proposal has{" "}
@@ -138,7 +133,7 @@ export const ResultExplanation = ({
               fontWeight: 700,
             }}
           >
-            {formatPrettyPercent(noWithVetoRatio.toNumber())}
+            {formatPrettyPercent(noWithVetoRatio)}
           </span>{" "}
           of the total votes, including &ldquo;Abstain&rdquo;, which exceeds the{" "}
           <span
@@ -146,14 +141,14 @@ export const ResultExplanation = ({
               fontWeight: 700,
             }}
           >
-            {formatPrettyPercent(vetoThreshold)} threshold
+            {formatPrettyPercent(vetoThreshold as Ratio<number>)} threshold
           </span>
           . If the proposal concludes with this voting outcome, it will be
           rejected regardless of &ldquo;Yes&rdquo; votes.
         </Text>
       );
 
-    if (yesRatio.lt(threshold))
+    if (yesNonRatio < threshold)
       return (
         <Text variant="body2">
           The proposal has{" "}
@@ -183,13 +178,13 @@ export const ResultExplanation = ({
     return (
       <Text variant="body2">
         Although the proposal successfully reached the voting quorum with a{" "}
-        {formatPrettyPercent(yesRatio.toNumber())} &ldquo;Yes&rdquo; rate, it
-        was not implemented due to technical reasons.
+        {formatPrettyPercent(yesNonRatio)} &ldquo;Yes&rdquo; rate, it was not
+        implemented due to technical reasons.
       </Text>
     );
 
   if (proposalData.status === ProposalStatus.REJECTED) {
-    if (yesRatio.gte(threshold))
+    if (yesNonRatio >= threshold)
       return (
         <Text variant="body2">
           This proposal did not meet the required quorum, resulting in its
@@ -197,7 +192,7 @@ export const ResultExplanation = ({
         </Text>
       );
 
-    if (noWithVetoRatio.gte(vetoThreshold))
+    if (noWithVetoRatio >= vetoThreshold)
       return (
         <Text variant="body2">
           The proposal is <Rejected /> due to either the quorum is not reached
@@ -207,7 +202,7 @@ export const ResultExplanation = ({
               fontWeight: 700,
             }}
           >
-            {formatPrettyPercent(noWithVetoRatio.toNumber())}{" "}
+            {formatPrettyPercent(noWithVetoRatio)}{" "}
           </span>
           of the total votes including &ldquo;Abstain&rdquo;, which exceeds the{" "}
           <span
@@ -215,7 +210,7 @@ export const ResultExplanation = ({
               fontWeight: 700,
             }}
           >
-            {formatPrettyPercent(vetoThreshold)} threshold
+            {formatPrettyPercent(vetoThreshold as Ratio<number>)} threshold
           </span>
           .
         </Text>
@@ -237,7 +232,7 @@ export const ResultExplanation = ({
           fontWeight: 700,
         }}
       >
-        {formatPrettyPercent(yesRatio.toNumber())} of &ldquo;Yes&rdquo;
+        {formatPrettyPercent(yesNonRatio)} of &ldquo;Yes&rdquo;
       </span>{" "}
       rate. As a result, the proposal has been passed, and its content will now
       be implemented.