+++ src/lib/pages/validators/components/validators-table/ValidatorsTableMobileCard.tsx
@@ -6,6 +6,7 @@ import { MobileCardTemplate, MobileLabel } from "lib/components/table";
 import { ValidatorBadge } from "lib/components/ValidatorBadge";
 import type {
   Option,
+  Ratio,
   Token,
   TokenWithValue,
   U,
@@ -76,7 +77,9 @@ export const ValidatorsTableMobileCard = ({
               <MobileLabel label="Voting Power" />
               <Text variant="body2" color="text.main">
                 {formatPrettyPercent(
-                  validator.votingPower.div(totalVotingPower).toNumber(),
+                  validator.votingPower
+                    .div(totalVotingPower)
+                    .toNumber() as Ratio<number>,
                   2,
                   true
                 )}
@@ -102,7 +105,11 @@ export const ValidatorsTableMobileCard = ({
                 color={isZeroUptime ? "error.main" : "text.main"}
                 fontWeight={isZeroUptime ? 700 : undefined}
               >
-                {formatPrettyPercent((validator.uptime ?? 0) / 100, 0, true)}
+                {formatPrettyPercent(
+                  ((validator.uptime ?? 0) / 100) as Ratio<number>,
+                  0,
+                  true
+                )}
               </Text>
             </Flex>
           </Flex>
@@ -113,7 +120,11 @@ export const ValidatorsTableMobileCard = ({
               color={isMinCommissionRate ? "success.main" : "text.main"}
               fontWeight={isMinCommissionRate ? 700 : undefined}
             >
-              {formatPrettyPercent(validator.commissionRate, 2, true)}
+              {formatPrettyPercent(
+                validator.commissionRate as Ratio<number>,
+                2,
+                true
+              )}
             </Text>
           </Flex>
         </Flex>